refactor file/folder/path getters
