+++ freemocap/gui/qt/widgets/active_recording_widget.py
@@ -9,7 +9,7 @@ from pyqtgraph.parametertree import Parameter, ParameterTree
 from freemocap.parameter_info_models.recording_info_model import (
     RecordingInfoModel,
 )
-from freemocap.system.paths_and_files_names import get_most_recent_recording_path
+from freemocap.system.paths_and_filenames.path_getters import get_most_recent_recording_path
 
 logger = logging.getLogger(__name__)
