+++ freemocap/gui/qt/widgets/directory_view_widget.py
@@ -8,7 +8,7 @@ from PyQt6.QtGui import QFileSystemModel
 from PyQt6.QtWidgets import QLabel, QMenu, QTreeView, QVBoxLayout, QWidget, QPushButton
 from qtpy import QtGui
 
-from freemocap.system.paths_and_files_names import get_recording_session_folder_path
+from freemocap.system.paths_and_filenames.path_getters import get_recording_session_folder_path
 from freemocap.system.open_file import open_file
 
 logger = logging.getLogger(__name__)
