+++ freemocap/tests/utilities/process_recording_headless.py
@@ -7,7 +7,7 @@ from freemocap.export_data.blender_stuff.export_to_blender import export_to_blen
 from freemocap.export_data.generate_jupyter_notebook.generate_jupyter_notebook import generate_jupyter_notebook
 from freemocap.parameter_info_models.recording_info_model import RecordingInfoModel
 from freemocap.parameter_info_models.recording_processing_parameter_models import RecordingProcessingParameterModel
-from freemocap.system.paths_and_files_names import get_blender_file_path
+from freemocap.system.paths_and_filenames.path_getters import get_blender_file_path
 
 logger = logging.getLogger(__name__)
